Created an error handler and an cog to customize the bot to your needs
