+++ base_folder/bot/config/Permissions.py
@@ -1,6 +1,8 @@
 import discord
 from discord.ext import commands
-from modules.base.db_management import get_settings_role, get_role
+from base_folder.bot.modules.base.db_management import get_settings_role, get_role
+
+# TODO: Rewrite the permission system so that it uses groups with permission instead of roles
 
 
 def is_dev():
