+++ base_folder/bot/modules/listener/listener_member.py
@@ -1,7 +1,7 @@
 import discord
 from discord.ext import commands
 from base_folder.bot.modules.base.db_management import is_user_indb, get_settings_role, get_leave_channel, \
-    get_role, on_error, get_prefix
+    get_role, get_prefix
 
 
 class ListenerMember(commands.Cog):
@@ -22,14 +22,6 @@ class ListenerMember(commands.Cog):
         channel_id = await get_leave_channel(member.guild.id)
         channel = member.guild.get_channel(channel_id)
         await channel.send('User {0.mention} left the server...'.format(member))
-
-    @commands.Cog.listener()
-    async def on_command_error(self, ctx, ex):
-        await on_error(ctx.guild.id, ex)
-        await ctx.channel.purge(limit=1)
-        error = get_prefix(ctx.guild.id)
-        await ctx.send(f"Please check with {error}"
-                       f" help the usage of this command or talk to your dev or admin.")
 
 
 def setup(client):
